pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lion"
version = "1.0.6"
description = "Linehaul Optimization for Network"
readme = "README.md"
requires-python = ">=3.12"
authors = [
  { name = "Alireza Asadi", email = "alireza.asadi@fedex.com" }
]
maintainers = [
  { name = "Alireza Asadi", email = "alireza.asadi@fedex.com" }
]
dependencies = [
  "flask~=3.1",
  "pandas==2.3.2",
  "numpy==2.3.2",
  "flask-bootstrap==3.3.7.1",
  "flask-sqlalchemy~=3.1",
  "flask-bcrypt~=1.0",
  "requests~=2.32",
  "cachetools~=6.1",
  "psutil==7.0.0",
  "openpyxl==3.1.5",
  "reportlab~=4.4",
  "gurobipy==12.0.3",
  "ortools==9.14.6206",
  "haversine~=2.9",
  "humanfriendly~=10.0",
  "waitress~=3.0",
  "setuptools~=80.0",
  "python-dotenv~=1.1",
  "pyodbc==5.2.0",
  "googletrans==4.0.2",
  "setproctitle~=1.3.7",
  "cryptography~=46.0.3",
  "pycparser~=2.23",
  "pyjwt~=2.10.1"
]


[project.scripts]
lion = "lion.__main__:main"

[tool.setuptools]
package-dir = { "" = "src" }
py-modules = ["dct_tour"] 
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["lion", "lion.*"]